requirements.txt
apify-client==1.6.3
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
google-genai==0.3.0
gspread==6.2.1

